requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Database and Vector Store
chromadb==0.4.24
chroma-hnswlib==0.7.3

# AI/ML Libraries
transformers==4.41.1
sentence-transformers==2.7.0
torch==2.3.1
numpy==1.24.3
scikit-learn==1.6.1
scipy==1.13.1
tokenizers==0.19.1
safetensors==0.6.2

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.1.0

# HTTP and API
requests==2.32.3
httpx==0.28.1
httpcore==1.0.9

# Data Processing
pydantic==2.11.7
pydantic_core==2.33.2
orjson==3.11.1
python-dateutil==2.9.0.post0

# Utilities
python-dotenv==1.0.1
click==8.1.8
typer==0.16.0
tqdm==4.67.1
rich==14.1.0
coloredlogs==15.0.1
humanfriendly==10.0

# File and System
fsspec==2025.7.0
filelock==3.18.0
pathlib2==2.3.7

# Security (Updated versions)
bcrypt==4.4.0
urllib3==2.2.0

# Networking
websockets==15.0.1
websocket-client==1.8.0

# Optional: Monitoring (remove if not needed)
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0

# Development and Testing (consider moving to requirements-dev.txt)
pytest==8.0.0
pytest-asyncio==0.24.0
mypy==1.10.0
black==24.10.0
flake8==7.1.1
loguru==0.7.2

# Dependencies (auto-generated, keep for compatibility)
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
backoff==2.2.1
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
flatbuffers==25.2.10
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.2.1
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
overrides==7.7.0
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyMuPDFb==1.23.7
PyPika==0.48.9
pyproject_hooks==1.2.0
PyYAML==6.0.2
regex==2025.7.34
requests-oauthlib==2.0.0
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvloop==0.21.0
watchfiles==1.1.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0